refactor: replace inline mock implementation of Select with centralized MockSelect from test-mocks
